+++ one/util.py
@@ -14,9 +14,12 @@ from packaging import version
 
 import one.alf.exceptions as alferr
 from one.alf.files import rel_path_parts, get_session_path, get_alf_path, remove_uuid_string
-from one.alf.spec import FILE_SPEC, regex as alf_regex
+from one.alf.spec import QC, FILE_SPEC, regex as alf_regex
 
 logger = logging.getLogger(__name__)
+
+QC_TYPE = pd.CategoricalDtype(categories=[e.name for e in sorted(QC)], ordered=True)
+"""pandas.api.types.CategoricalDtype: The cache table QC column data type."""
 
 
 def Listable(t):
@@ -51,21 +54,21 @@ def ses2records(ses: dict):
 
     # Extract datasets table
     def _to_record(d):
-        rec = dict(file_size=d['file_size'], hash=d['hash'], exists=True)
-        rec['id'] = d['id']
+        rec = dict(file_size=d['file_size'], hash=d['hash'], exists=True, id=d['id'])
         rec['eid'] = session.name
         file_path = urllib.parse.urlsplit(d['data_url'], allow_fragments=False).path.strip('/')
         file_path = get_alf_path(remove_uuid_string(file_path))
         rec['session_path'] = get_session_path(file_path).as_posix()
         rec['rel_path'] = file_path[len(rec['session_path']):].strip('/')
         rec['default_revision'] = d['default_revision'] == 'True'
+        rec['qc'] = d.get('qc', 'NOT_SET')
         return rec
 
     if not ses.get('data_dataset_session_related'):
         return session, pd.DataFrame()
     records = map(_to_record, ses['data_dataset_session_related'])
     index = ['eid', 'id']
-    datasets = pd.DataFrame(records).set_index(index).sort_index()
+    datasets = pd.DataFrame(records).set_index(index).sort_index().astype({'qc': QC_TYPE})
     return session, datasets
 
 
@@ -106,15 +109,16 @@ def datasets2records(datasets, additional=None) -> pd.DataFrame:
             rec['session_path'] = rec['session_path'].as_posix()
         rec['rel_path'] = file_path[len(rec['session_path']):].strip('/')
         rec['default_revision'] = d['default_dataset']
+        rec['qc'] = d.get('qc')
         for field in additional or []:
             rec[field] = d.get(field)
         records.append(rec)
 
     index = ['eid', 'id']
     if not records:
-        keys = (*index, 'file_size', 'hash', 'session_path', 'rel_path', 'default_revision')
+        keys = (*index, 'file_size', 'hash', 'session_path', 'rel_path', 'default_revision', 'qc')
         return pd.DataFrame(columns=keys).set_index(index)
-    return pd.DataFrame(records).set_index(index).sort_index()
+    return pd.DataFrame(records).set_index(index).sort_index().astype({'qc': QC_TYPE})
 
 
 def parse_id(method):
@@ -148,9 +152,7 @@ def parse_id(method):
 
 
 def refresh(method):
-    """
-    Refresh cache depending of query_type kwarg.
-    """
+    """Refresh cache depending on query_type kwarg."""
 
     @wraps(method)
     def wrapper(self, *args, **kwargs):
@@ -283,8 +285,9 @@ def _file_spec(**kwargs):
     return filespec
 
 
-def filter_datasets(all_datasets, filename=None, collection=None, revision=None,
-                    revision_last_before=True, assert_unique=True, wildcards=False):
+def filter_datasets(
+        all_datasets, filename=None, collection=None, revision=None, revision_last_before=True,
+        qc=QC.FAIL, ignore_qc_not_set=False, assert_unique=True, wildcards=False):
     """
     Filter the datasets cache table by the relative path (dataset name, collection and revision).
     When None is passed, all values will match.  To match on empty parts, use an empty string.
@@ -305,6 +308,11 @@ def filter_datasets(all_datasets, filename=None, collection=None, revision=None,
         When true and no exact match exists, the (lexicographically) previous revision is used
         instead.  When false the revision string is matched like collection and filename,
         with regular expressions permitted.
+    qc : str, int, one.alf.spec.QC
+        Returns datasets at or below this QC level.  Integer values should correspond to the QC
+        enumeration NOT the qc category column codes in the pandas table.
+    ignore_qc_not_set : bool
+        When true, do not return datasets for which QC is NOT_SET.
     assert_unique : bool
         When true an error is raised if multiple collections or datasets are found.
     wildcards : bool
@@ -333,6 +341,14 @@ def filter_datasets(all_datasets, filename=None, collection=None, revision=None,
     Filter by filename parts
 
     >>> datasets = filter_datasets(all_datasets, dict(object='spikes', attribute='times'))
+
+    Filter by QC outcome - datasets with WARNING or better
+
+    >>> datasets filter_datasets(all_datasets, qc='WARNING')
+
+    Filter by QC outcome and ignore datasets with unset QC - datasets with PASS only
+
+    >>> datasets filter_datasets(all_datasets, qc='PASS', ignore_qc_not_set=True)
 
     Notes
     -----
@@ -368,7 +384,16 @@ def filter_datasets(all_datasets, filename=None, collection=None, revision=None,
 
     # Build regex string
     pattern = alf_regex('^' + spec_str, **regex_args)
-    match = all_datasets[all_datasets['rel_path'].str.match(pattern)]
+    path_match = all_datasets['rel_path'].str.match(pattern)
+
+    # Test on QC outcome
+    qc = QC.validate(qc)
+    qc_match = all_datasets['qc'].le(qc.name)
+    if ignore_qc_not_set:
+        qc_match &= all_datasets['qc'].ne('NOT_SET')
+
+    # Filter datasets on path and QC
+    match = all_datasets[path_match & qc_match]
     if len(match) == 0 or not (revision_last_before or assert_unique):
         return match
 
@@ -558,7 +583,7 @@ def cache_int2str(table: pd.DataFrame) -> pd.DataFrame:
     return table
 
 
-def patch_cache(table: pd.DataFrame, min_api_version=None) -> pd.DataFrame:
+def patch_cache(table: pd.DataFrame, min_api_version=None, name=None) -> pd.DataFrame:
     """Reformat older cache tables to comply with this version of ONE.
 
     Currently this function will 1. convert integer UUIDs to string UUIDs; 2. rename the 'project'
@@ -570,10 +595,15 @@ def patch_cache(table: pd.DataFrame, min_api_version=None) -> pd.DataFrame:
         A cache table (from One._cache).
     min_api_version : str
         The minimum API version supported by this cache table.
+    name : {'dataset', 'session'} str
+        The name of the table.
     """
     min_version = version.parse(min_api_version or '0.0.0')
     table = cache_int2str(table)
     # Rename project column
     if min_version < version.Version('1.13.0') and 'project' in table.columns:
         table.rename(columns={'project': 'projects'}, inplace=True)
+    if name == 'datasets' and min_version < version.Version('2.7.0') and 'qc' not in table.columns:
+        qc = pd.Categorical.from_codes(np.zeros(len(table.index), dtype=int), dtype=QC_TYPE)
+        table = table.assign(qc=qc)
     return table